requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asyncio==3.4.3
certifi==2024.2.2; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
deprecation==2.1.0
fastapi==0.110.2; python_version >= '3.8'
gotrue==2.4.2; python_version >= '3.8' and python_version < '4.0'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
loguru==0.7.2; python_version >= '3.5'
packaging==24.0; python_version >= '3.7'
postgrest==0.16.3; python_version >= '3.8' and python_version < '4.0'
pydantic==2.7.0; python_version >= '3.8'
pydantic-core==2.18.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
realtime==1.0.4; python_version >= '3.8' and python_version < '4.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slack-bolt==1.18.1; python_version >= '3.6'
slack-sdk==3.27.1; python_version >= '3.6'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
storage3==0.7.4; python_version >= '3.8' and python_version < '4.0'
strenum==0.4.15
supabase==2.4.3; python_version >= '3.8' and python_version < '4.0'
supafunc==0.4.5; python_version >= '3.8' and python_version < '4.0'
typing-extensions==4.11.0; python_version >= '3.8'
uvicorn==0.29.0; python_version >= '3.8'
websockets==12.0; python_version >= '3.8'
